refactor(rules): share state control flow helpers
